Remove duplication in flags, undefined in ROOTAnalGas class and other minor corrections.
Several flags were duplicated in ActarSimROOTAnalGas and their value was not properly set by the Messenger. Now the flags
are only set in ActarSimROOTAnalysis class and the value used (via the global pointer) in other daughter classes.
Also it corrects a duplication in the primaryInfo TClonesArray in the TTree, which now is solved by calling a Clear()
just after filling the The_ACTAR_Event_Tree TTree.