First attemp for a doxygen documentation of the code.
A complete doxygen documentation is included. The code has been reviewed to uniformize the style and
remove irrelevant comments. Please, stick to the used style of the C++, including a clear and strict
indentation and removal of spaces in blanck lines or after each code line (as "whitespace" package makes
in atom editor by default).