Modifications to work with ROOT6 and Geant4.10.02
The modifications include: using CLHEP namespace when required for units (to avoid shadowing with other variables) or
load the header files required in Geant4.10 for units when there is no collision. Removal of some unused ROOT header files.
Few modifications to solve declaring several times a variable (shadowing warnings). Small modifications on the name of
some physics model as reported in Geant4.10 release notes. Some work on doxygen too, moving it
out of CMakeLists.txt to make it independently, update the version and move to a different output directory not linked with
any account.