Include default values for Flags in the constructor.
Rewrite energy deposited in gas and silicon (old GetDeltaEnergy).

To avoid warning messages about removing GetDeltaEnergy, the behaviour of the
function is directly implemented in the sensitive detectors code. Additionally the
CH4 default gas has been renamed to CH4_STP to avoid confusion.

Substitution of old gas definition in macros.

Add independent datamember in beamInfo for next event Z vertex position limit.

Solves a bug in the code. The zVertex position from the beamInfo was mixing information from the next event, as
the limit for the next event was calculated and saved in the zVertex element before the TClonesArray was filled.
Therefore, there was an event mixing, now solved.