Implement PAI physics model for testing energy loss in gases
The PAI has been recommended (Geant4 physics list, RD51, ...) for describing the energy loss in gases
at low energies. The model has been included in the "activeGas" region, defined in the logical gas
volume. The code copies from testEm8 in some parts. Additionally the commit includes correction in the
Z number assigned to the Argon and on the input arguments in readerPads.C.